app/requirements.txt
fastapi
httpx
python-dotenv
uvicorn
pyjwt